# pizzaSiteAppModules/api/src/main/resources/openapi/new/0420.yaml
openapi: 3.0.0
info:
  title: Pizza project API
  version: 1.1.0
  description: |
    Pizza API allows you to remotely manage the pizzeria.
    ### Description of the project and requirements:
    * Authorization is required to work with the API. Authorization is done by the bearer-token 😎
servers:
  - url: http://localhost:8090/
paths:
...
  /api/base/all:
    get:
      summary: Get all bases
      tags:
        - Bases
      operationId: get all bases
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Base'
        '403':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
  /api/base/{id}:
    get:
      summary: Get base by id
      description: Get base
      operationId: get one base
      tags:
        - Bases
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200' :
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Base'
        '403':
          description: Access denied
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Only admins are allowed access"
        '404':
          description: Incorrect data
          content:
            appllication/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The data entered is not correct"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      in: path
      name: id
      required: true
      description: record id
      schema:
        type: integer
...
  schemas:
...
    Base:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Thin crust'
        size:
          type: string
          example: 'Small'
        price:
          type: number
          example: '3.50'
...
security:
  - bearerAuth: []